move paddock stuff to website
