+++ components/paddock/b4mad_racing_website/urls.py
@@ -17,4 +17,18 @@ urlpatterns = [
         name="profile-subscriptions",
     ),
     path("profile/", views.ProfileRedirectView.as_view(), name="profile-redirector"),
+    # path('pitcrew', TemplateView.as_view(template_name='pitcrew.html'), name="pitcrew"),
+    # path("fastlap/<str:game>/<str:track>/<str:car>", views.fastlap_index, name="fastlap_index"),
+    # path("fastlap/<str:game>/<str:track>>", views.fastlap_index, name="fastlap_index"),
+    # path("fastlap/<str:game>/", views.fastlap_index, name="fastlap_index"),
+    # path("fastlap/<int:fastlap_id>", views.fastlap_view, name="fastlap"),
+    path("fastlap/", views.fastlap, name="fastlap"),
+    # path("fastlap_data/<int:fastlap_id>", views.fastlap_data, name="fastlap_data"),
+    # path("fastlap/", views.fastlap_index, name="fastlap_index"),
+    # path("pitcrew/<str:driver_name>", views.pitcrew_view, name="pitcrew"),
+    # path("pitcrew/", views.pitcrew_index, name="pitcrew_index"),
+    # path("coach/", CoachView.as_view(), name="coach"),
+    path("session/<int:session_id>", views.session, name="session"),
+    path("session/<int:session_id>/<int:lap>", views.session, name="session"),
+    path("sessions/<int:game_id>/<int:car_id>/<int:track_id>", views.sessions, name="sessions"),
 ]
