--- components/paddock/paddock/__init__.py
@@ -1,15 +1,0 @@
-# import toml
-
-# from .celery import app as celery_app
-
-
-# def get_version_from_pyproject():
-#     with open("pyproject.toml", "r") as file:
-#         pyproject = toml.load(file)
-#         return pyproject["project"]["version"]
-
-
-# __version__ = get_version_from_pyproject()
-
-
-# __all__ = ("celery_app",)
