+++ components/paddock/paddock/urls.py
@@ -8,30 +8,10 @@ from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
 
-from . import views
-
-# from .views import CoachView
-
 urlpatterns = [
-    # path("", views.index, name="home"),
     path("", include("b4mad_racing_website.urls")),
     path("", include("django_prometheus.urls")),
     path("", include("api.urls")),
-    # path("", TemplateView.as_view(template_name="home.html"), name="home"),
-    # path('pitcrew', TemplateView.as_view(template_name='pitcrew.html'), name="pitcrew"),
-    # path("fastlap/<str:game>/<str:track>/<str:car>", views.fastlap_index, name="fastlap_index"),
-    # path("fastlap/<str:game>/<str:track>>", views.fastlap_index, name="fastlap_index"),
-    # path("fastlap/<str:game>/", views.fastlap_index, name="fastlap_index"),
-    # path("fastlap/<int:fastlap_id>", views.fastlap_view, name="fastlap"),
-    path("fastlap/", views.fastlap, name="fastlap"),
-    # path("fastlap_data/<int:fastlap_id>", views.fastlap_data, name="fastlap_data"),
-    # path("fastlap/", views.fastlap_index, name="fastlap_index"),
-    # path("pitcrew/<str:driver_name>", views.pitcrew_view, name="pitcrew"),
-    # path("pitcrew/", views.pitcrew_index, name="pitcrew_index"),
-    # path("coach/", CoachView.as_view(), name="coach"),
-    path("session/<int:session_id>", views.session, name="session"),
-    path("session/<int:session_id>/<int:lap>", views.session, name="session"),
-    path("sessions/<int:game_id>/<int:car_id>/<int:track_id>", views.sessions, name="sessions"),
     path("admin/", admin.site.urls),
     path("accounts/", include("allauth.urls")),
     path("explorer/", include("explorer.urls")),
